src/backend/workflow23/requirements.txt
asgiref==3.6.0
black==23.1.0
click==8.1.3
Django==4.1.6
django-environ==0.9.0
django-js-asset==2.0.0
django-mptt==0.14.0
gunicorn==20.1.0
mypy-extensions==0.4.3
packaging==23.0
pathspec==0.11.0
platformdirs==2.6.2
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
